requirements.txt
# python==3.12
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
pydantic==2.5.0
pandas
